requirements.txt
aiohttp==3.7.3
APScheduler==3.6.3
async-lru==1.0.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
ffmpeg==1.4
gevent==20.9.0
greenlet==0.4.17
grequests==0.6.0
idna==2.10
lxml==4.6.1
multidict==5.0.2
peewee==3.14.0
pyaes==1.6.1
pycparser==2.20
Pyrogram==1.0.7
PySocks==1.7.1
python-decouple==3.3
python-telegram-bot==13.0
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
TgCrypto==1.2.2
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.2.0
